Move MetdataReferenceEqualityComparer to OmniSharp.Roslyn/Utilities
